Break the sql abstraction layer apart and add unit tests.
